# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'Autpost API'
  description: 'API documentation for the Autpost application'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:9527/api'
    description: 'Local development server'
paths:
...
  /todos:
    get:
      tags:
        - Todos
      summary: 'Get all todos for the authenticated user'
      operationId: 18e676fc7a1838c48c63274e52066ac8
      parameters:
        -
          name: states
          in: query
          required: false
          schema:
            description: 'Comma-separated states to filter by (e.g., pending,in_progress)'
            type: string
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'List of todos with pagination'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Todos
      summary: 'Create a new todo'
      operationId: ae1281faa0d4aef5c4f3e277c6a57e7c
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
              properties:
                recurring_todo_id:
                  type: integer
                  nullable: true
                title:
                  type: string
                  maxLength: 25
                note:
                  type: string
                  maxLength: 25
                  nullable: true
                due_time:
                  type: string
                  format: date-time
                  nullable: true
                is_whole_day:
                  type: boolean
                  nullable: true
                state:
                  type: string
                  enum:
                    - pending
                    - in_progress
                    - completed
                    - cancelled
                  nullable: true
              type: object
      responses:
        201:
          description: 'Todo created successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  '/todos/{id}':
    get:
      tags:
        - Todos
      summary: 'Get a single todo'
      operationId: 1b99eaebc4bafa42301fbf780b77da93
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo details'
        404:
          description: 'Todo not found'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Todos
      summary: 'Update a todo'
      operationId: c10cfb143d3c35afe02449b07608baba
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  maxLength: 255
                note:
                  type: string
                  nullable: true
                due_time:
                  type: string
                  format: date-time
                  nullable: true
                is_whole_day:
                  type: boolean
                  nullable: true
                state:
                  type: string
                  enum:
                    - pending
                    - in_progress
                    - completed
                    - cancelled
                  nullable: true
              type: object
      responses:
        200:
          description: 'Todo updated successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Todo not found'
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Todos
      summary: 'Soft delete a todo'
      operationId: e6e24dd784980cf6f9fe389df03fdd14
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo deleted successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Todo not found'
      security:
        -
          bearerAuth: []
  '/todos/{id}/restore':
    post:
      tags:
        - Todos
      summary: 'Restore a soft-deleted todo'
      operationId: fee52d138ce5c001ce166e42d3941ff3
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo restored successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Todo not found'
      security:
        -
          bearerAuth: []
  '/todos/{id}/force':
    delete:
      tags:
        - Todos
      summary: 'Permanently delete a todo'
      operationId: b0f84fca5128e2aee82ad6d768ac2233
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo permanently deleted'
        401:
          description: Unauthenticated
        404:
          description: 'Todo not found'
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter your Sanctum token'
      bearerFormat: token
      scheme: bearer
...